feat: policy criada com todos os filtros necessários até então, mudançãs e pequenas coreções no controller e adição de query no restaurante respository.

+++ app/Domains/Restaurante/Controllers/RestauranteController.php
@@ -2,12 +2,12 @@
 
 namespace App\Domains\Restaurante\Controllers;
 
-use App\Domains\Catalogo\Resources\CategoriaResource;
 use App\Domains\Restaurante\Requests\Restaurante\StoreRestauranteRequest;
 use App\Domains\Restaurante\Requests\Restaurante\UpdateRestauranteRequest;
 use App\Domains\Restaurante\Resources\RestauranteResource;
 use App\Domains\Restaurante\Services\RestauranteService;
 use App\Http\Controllers\Controller;
+use App\Models\Restaurante;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 
@@ -21,29 +21,35 @@ class RestauranteController extends Controller
 
     public function index(): AnonymousResourceCollection
     {
-        return RestauranteResource::collection($this->restauranteService->findAll());
+        $this->authorize('viewAny', Restaurante::class);
+        return RestauranteResource::collection($this->restauranteService->findAll(auth()->user()));
     }
 
     public function show(int $id): RestauranteResource
     {
         $restaurante = $this->restauranteService->find($id);
+        $this->authorize('view', $restaurante);
         return new RestauranteResource($restaurante);
     }
 
     public function store(StoreRestauranteRequest $request): RestauranteResource
     {
         $restaurante = $this->restauranteService->create($request->validated());
+        $this->authorize('create', Restaurante::class);
         return new RestauranteResource($restaurante);
     }
 
     public function update(UpdateRestauranteRequest $request, $id) : RestauranteResource
     {
         $restaurante = $this->restauranteService->update($id, $request->validated());
+        $this->authorize('update', $restaurante);
         return new RestauranteResource($restaurante);
     }
 
     public function destroy(int $id) : JsonResponse
     {
+        $restaurante = $this->restauranteService->find($id);
+        $this->authorize('delete', $restaurante);
         $this->restauranteService->delete($id);
         return response()->json(['message' => 'Restaurante excluido com sucesso!']);
     }
